+++ client/src/pages/Login.js
@@ -29,16 +29,15 @@ function Login(props) {
     });
   };
   return (
-  <div class="modal modal-signin position-static d-block bg-secondary py-5" tabindex="-1" role="dialog" id="modalSignin">
-    <div class="modal-dialog" role="document">
-      <div class="modal-content rounded-5 shadow">
-        <div class="modal-header p-5 pb-4 border-bottom-0">
-          <h2 class="fw-bold mb-0">Sign up for free</h2>
-          <button type="button" class="btn-close" data-bs-dismiss="modal" aria-label="Close"></button>
-        </div>
-        <div class="modal-body p-5 pt-0">
+  <div className="modal modal-signin position-static d-block  bg-secondary py-5" tabIndex="-1" role="dialog" id="modalSignin">
+    <div className="modal-dialog" role="document">
+      <div className="modal-content rounded-5 shadow">
+        <div className="modal-header p-5 pb-4 border-bottom-0">
+          <h2 className="fw-bold mb-0 text-centered">Login</h2>
+        </div>      
+        <div className="modal-body p-5 pt-0">
           <form onSubmit={handleFormSubmit}>
-              <div class="form-floating mb-3">
+              <div className="form-floating mb-3">
                 <input
                   className="form-control rounded-4"
                   placeholder="youremail@test.com"
@@ -47,9 +46,9 @@ function Login(props) {
                   id="email"
                   onChange={handleChange}
                 />
-                <label htmlfor="floatingEmail">Email:</label>
+                <label htmlFor="floatingEmail">Email:</label>
           </div>
-          <div class="form-floating mb-3">
+          <div className="form-floating mb-3">
             <input
               className="form-control rounded-4"
               placeholder="******"
@@ -58,10 +57,10 @@ function Login(props) {
               id="pwd"
               onChange={handleChange}
             />
-            <label htmlfor="floatingPassword">Password</label>
+            <label htmlFor="floatingPassword">Password</label>
           </div>
-          <button class="w-100 mb-2 btn btn-lg rounded-4 btn-primary" type="submit">Submit</button>
-         </form>
+          <button className="w-100 mb-2 btn btn-lg rounded-4 btn-primary" type="submit">Submit</button>
+        </form>
         <Link to="/signup">← Go to Signup</Link>
       </div>
     </div>
